restructure widget tree
